requirements.txt
flask
openai
requests
python-dotenv
markdown
pyjwt
cryptography
